fix: correctly setup multiple PPAs

+++ chimg/chroot.py
@@ -482,24 +482,29 @@ GRUB_FORCE_PARTUUID={partuuid}"""
         Setup all configured PPAs
         """
         if len(self._ctx.conf["ppas"]) > 0:
-            for ppa in self._ctx.conf["ppas"]:
-                with self._ppa_setup(
-                    ppa["name"],
-                    ppa["uri"],
-                    ppa["suites"],
-                    ppa["components"],
-                    ppa["fingerprint"],
-                    ppa["signed_by"],
-                    ppa["username"],
-                    ppa["password"],
-                    ppa["auth_lines"],
-                    ppa["pin_name"],
-                    ppa["pin_priority"],
-                ):
-                    cmd = ["/usr/sbin/chroot", self._ctx.chroot_path, "apt-cache", "policy"]
-                    out, err = run_command(cmd)
-                    logger.info(out)
-                    yield
+            with ExitStack() as stack:
+                for ppa in self._ctx.conf["ppas"]:
+                    stack.enter_context(
+                        self._ppa_setup(
+                            ppa["name"],
+                            ppa["uri"],
+                            ppa["suites"],
+                            ppa["components"],
+                            ppa["keep"],
+                            ppa["fingerprint"],
+                            ppa["signed_by"],
+                            ppa["username"],
+                            ppa["password"],
+                            ppa["auth_lines"],
+                            ppa["pin_name"],
+                            ppa["pin_priority"],
+                        )
+                    )
+                logger.info("All PPAs setup")
+                cmd = ["/usr/sbin/chroot", self._ctx.chroot_path, "apt-cache", "policy"]
+                out, err = run_command(cmd)
+                logger.info(out)
+                yield
         else:
             # no PPAs setup - but do at least one apt-get update
             self._apt_update()
@@ -512,6 +517,7 @@ GRUB_FORCE_PARTUUID={partuuid}"""
         repo_uri: str,
         repo_suites: List[str],
         repo_components: List[str],
+        keep: bool,
         repo_key_fingerprint: Optional[str] = None,
         signed_by: Optional[str] = None,
         repo_username: Optional[str] = None,
@@ -572,17 +578,18 @@ Pin-Priority: {repo_pin_priority}
         logger.info("PPA added")
         yield
         # cleanup the PPA
-        logger.info("Removing PPA ...")
-        if os.path.exists(f"{self._ctx.chroot_path}/etc/apt/sources.list.d/{name}.sources"):
-            os.remove(f"{self._ctx.chroot_path}/etc/apt/sources.list.d/{name}.sources")
-        if os.path.exists(f"{self._ctx.chroot_path}/etc/apt/trusted.gpg.d/{name}.gpg"):
-            os.remove(f"{self._ctx.chroot_path}/etc/apt/trusted.gpg.d/{name}.gpg")
-        if os.path.exists(f"{self._ctx.chroot_path}/etc/apt/auth.conf.d/{name}.conf"):
-            os.remove(f"{self._ctx.chroot_path}/etc/apt/auth.conf.d/{name}.conf")
-        if os.path.exists(f"{self._ctx.chroot_path}/etc/apt/preferences.d/{name}.pref"):
-            os.remove(f"{self._ctx.chroot_path}/etc/apt/preferences.d/{name}.pref")
-        self._apt_update()
-        logger.info("PPA removed")
+        if not keep:
+            logger.info("Removing PPA ...")
+            if os.path.exists(f"{self._ctx.chroot_path}/etc/apt/sources.list.d/{name}.sources"):
+                os.remove(f"{self._ctx.chroot_path}/etc/apt/sources.list.d/{name}.sources")
+            if os.path.exists(f"{self._ctx.chroot_path}/etc/apt/trusted.gpg.d/{name}.gpg"):
+                os.remove(f"{self._ctx.chroot_path}/etc/apt/trusted.gpg.d/{name}.gpg")
+            if os.path.exists(f"{self._ctx.chroot_path}/etc/apt/auth.conf.d/{name}.conf"):
+                os.remove(f"{self._ctx.chroot_path}/etc/apt/auth.conf.d/{name}.conf")
+            if os.path.exists(f"{self._ctx.chroot_path}/etc/apt/preferences.d/{name}.pref"):
+                os.remove(f"{self._ctx.chroot_path}/etc/apt/preferences.d/{name}.pref")
+            self._apt_update()
+            logger.info("PPA removed")
 
     @contextmanager
     def _policy_rc_runlevel_ops(self):
